backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
flask-session==0.5.0
authlib==1.2.1
python-dotenv==1.0.0
requests==2.31.0
pymysql==1.1.0
numpy==1.24.3
groq==0.4.2
python-jose==3.3.0
cryptography==41.0.3
